// web/src/generated/StatusAccount.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SyncLog } from "./SyncLog";
import type { SyncProgress } from "./SyncProgress";

export type StatusAccount = { id: number, name: string, kind: string, host: string, username: string, message_count: number, index_size_bytes: number, last_sync_at: number | null, last_sync_status: string | null, sync_interval_mins: number, excluded_folders: Array<string>, progress: SyncProgress | null, recent_syncs: Array<SyncLog>, };